Support u64 and u32 in addition of u128

+++ tests/compile/fail/invalid-struct-primitive.rs
@@ -12,6 +12,6 @@ mod mod_id {
 
 #[derive(Identifier, Eq, PartialEq)]
 #[identifier(with = "mod_id")]
-struct Id(u64);
+struct Id(u16);
 
 fn main() {}
